Refactor common implementation
