+++ messages.json
@@ -9,14 +9,14 @@
   "NON_CIVILIAN_WIN": "There's only one Civilian left, **Undercovers and Mr. White win**",
   "MR_WHITE_WIN": "There's no difficulty guessing the word, **Mr.White win**",
   "SUMMARY": "**Civilians** ({civilian_word}): {civilians}\n**Undercovers** ({undercover_word}): {undercovers}\n**Mr. White**: {mr_whites}",
-  "HOW_TO": "```How to Play\n0. Check your DM\n1. Remember and don't share your word to other players\n2. Mr. White won't get any word\n3. All players won't know others' role\n4. Describe your word by following the order\n5. Each player must describe their words\n6. Choose someone to be eliminated\n7. The chosen one must type !eliminated to eliminate himself\n8. Repeat step 1-7 each turn\n9. The game will end if meets one of the conditions below\n\nWin Conditions\n0. Mr. White guess the Civilian's word correctly (Mr. White win)\n1. Undercovers and Mr. White are eliminated (Civilians win)\n2. There is only one civilian left (Undercovers and Mr. White Win)```",
+  "HOW_TO_COMMAND": "```How to Play\n0. Check your DM\n1. Remember and don't share your word to other players\n2. Mr. White won't get any word\n3. All players won't know others' role\n4. Describe your word by following the order\n5. Each player must describe their words\n6. Choose someone to be eliminated\n7. The chosen one must type !eliminated to eliminate himself\n8. Repeat step 1-7 each turn\n9. The game will end if meets one of the conditions below\n\nWin Conditions\n0. Mr. White guess the Civilian's word correctly (Mr. White win)\n1. Undercovers and Mr. White are eliminated (Civilians win)\n2. There is only one civilian left (Undercovers and Mr. White Win)```",
   "POLL_DECIDED": "Poll decided, {player} will be eliminated",
 
   "GUILD_ONLY_COMMAND": "You must run this command in a server",
   "DM_ONLY_COMMAND": "You can only run this command from private message",
-  "HUMAN_PLAYER_ONLY": "Sorry, bots are not allowed to join the game",
-  "NO_MULTIPLE_VOTES": "Sorry, you can only vote one player",
-  "SPECIFY_USER": "Please specify the user you want to apply the command to",
+  "BOT_PLAYER_FOUND": "Sorry, bots are not allowed to join the game",
+  "MULTIPLE_VOTES_FOUND": "Sorry, you can only vote one player",
+  "EMPTY_VOTE_FOUND": "Please specify the user you want to vote",
   "INVALID_PLAYER_NUMBER": "The number of players should be {min_player} up to {max_player}",
   "ONGOING_GAME_FOUND": "There's an ongoing game in this server, finish it first",
   "ONGOING_POLL_FOUND": "There's an ongoing poll in this channel, wait it to finish first",
@@ -26,8 +26,8 @@
   "PLAYER_NOT_FOUND": "Command failed: {player} is not playing in this channel",
   "PLAYER_ALREADY_KILLED": "Command failed: {player} has been eliminated before",
   "NOT_IN_GUESSING_TURN": "It's not yet the time for guessing",
-  "NO_VOTES_SUBMITTED": "No votes submitted. No players will be eliminated",
-  "NOT_ENOUGH_VOTES": "No more than half players voted. No players will be eliminated",
-  "MULTIPLE_PLAYERS_VOTED": "Multiple players voted. No players will be eliminated",
-  "VOTE_EXISTS": "{player} has voted before. You can only vote once"
+  "NO_VOTES_SUBMITTED": "No votes submitted, none will be eliminated",
+  "NOT_ENOUGH_VOTES": "No more than half players voted, none will be eliminated",
+  "MULTIPLE_PLAYERS_VOTED": "Multiple players voted, none will be eliminated",
+  "VOTE_EXISTS": "{player} has voted before, you can only vote once"
 }
